# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "BuilderRun": {
        "properties": {
          "completedAt": {
            "format": "date-time",
            "type": [
              "null",
              "string"
            ]
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "error": {
            "type": [
              "null",
              "string"
            ]
          },
          "familyId": {
            "format": "uuid",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "inputHash": {
            "type": "string"
          },
          "inputVersion": {
            "type": "string"
          },
          "libraryReleaseId": {
            "format": "uuid",
            "type": [
              "null",
              "string"
            ]
          },
          "model": {
            "type": "string"
          },
          "nextAttemptAt": {
            "format": "date-time",
            "type": [
              "null",
              "string"
            ]
          },
          "promptVersion": {
            "type": "string"
          },
          "provider": {
            "type": "string"
          },
          "retries": {
            "format": "int32",
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ]
          },
          "retryRound": {
            "format": "int32",
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ]
          },
          "sourceId": {
            "format": "uuid",
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "type": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "Candidate": {
        "properties": {
          "behavior": {
            "type": "string"
          },
          "boundary": {
            "type": "string"
          },
          "chunkId": {
            "format": "uuid",
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "createdDraftId": {
            "format": "uuid",
            "type": [
              "null",
              "string"
            ]
          },
          "createdKCId": {
            "format": "uuid",
            "type": [
              "null",
              "string"
            ]
          },
          "decision": {
            "type": "string"
          },
          "existingKCId": {
            "format": "uuid",
            "type": [
              "null",
              "string"
            ]
          },
          "familyId": {
            "format": "uuid",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "matches": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "quote": {
            "type": "string"
          },
          "reviewReason": {
            "type": "string"
          },
          "reviewedAt": {
            "format": "date-time",
            "type": [
              "null",
              "string"
            ]
          },
          "reviewedBy": {
            "format": "uuid",
            "type": [
              "null",
              "string"
            ]
          },
          "runId": {
            "format": "uuid",
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "suggestedAction": {
            "type": "string"
          },
          "type": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "DecisionInput": {
        "properties": {
          "behavior": {
            "type": [
              "null",
              "string"
            ]
          },
          "boundary": {
            "type": [
              "null",
              "string"
            ]
          },
          "decision": {
            "type": "string"
          },
          "existingKCId": {
            "format": "uuid",
            "type": [
              "null",
              "string"
            ]
          },
          "name": {
            "type": [
              "null",
              "string"
            ]
          },
          "reason": {
            "type": "string"
          }
        },
        "required": [
          "decision",
          "reason"
        ],
        "type": "object"
      },
...
      "ReasonInput": {
        "properties": {
          "reason": {
            "type": "string"
          }
        },
        "required": [
          "reason"
        ],
        "type": "object"
      },
...
      "RunInput": {
        "properties": {
          "provider": {
            "default": "Mock",
            "type": "string"
          },
          "sourceId": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "sourceId"
        ],
        "type": "object"
      },
      "Source": {
        "properties": {
          "allowExternalAI": {
            "type": "boolean"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "familyId": {
            "format": "uuid",
            "type": "string"
          },
          "fileId": {
            "format": "uuid",
            "type": [
              "null",
              "string"
            ]
          },
          "hash": {
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "text": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "usageScope": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "SourceInput": {
        "properties": {
          "allowExternalAI": {
            "default": false,
            "type": "boolean"
          },
          "text": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "usageScope": {
            "default": "FamilyOnly",
            "type": "string"
          }
        },
        "required": [
          "title",
          "text"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Learning.Api | v1",
    "version": "1.0.0"
  },
  "openapi": "3.1.1",
  "paths": {
...
    "/api/v1/builder/candidates/{id}:decide": {
      "post": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecisionInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Candidate"
                }
              }
            }
          }
        },
        "tags": [
          "Learning.Api"
        ]
      }
    },
    "/api/v1/builder/runs": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BuilderRun"
                }
              }
            }
          },
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BuilderRun"
                }
              }
            }
          }
        },
        "tags": [
          "Learning.Api"
        ]
      }
    },
    "/api/v1/builder/runs/{id}:retry": {
      "post": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReasonInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BuilderRun"
                }
              }
            }
          }
        },
        "tags": [
          "Learning.Api"
        ]
      }
    },
...
    "/api/v1/content/sources": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SourceInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Source"
                }
              }
            }
          },
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Source"
                }
              }
            }
          }
        },
        "tags": [
          "Learning.Api"
        ]
      }
    },
...
  },
...
}